refactor(trainer): extract Gen 9 sections + group fields into headed sections
Splits the SAV9SV and SAV9ZA conditional blocks out of TrainerInfoTab
into dedicated per-game components under
Pkmds.Rcl/Components/MainTabPages/Trainer/. Inside each section, fields
are now organized under .form-section-header rows (Identity & Records,
Location, Currency, Unlocks & Collections, Editors) instead of being
intermixed with action buttons in one undifferentiated grid.

This is the first step toward replacing the long @if (saveFile is X)
ladder in TrainerInfoTab with per-game section components — future PRs
can extract Sav8SwSh, Sav8La, Sav7, etc. the same way.

- New: TrainerInfoSav9SvSection (.razor + .razor.cs) — owns birthday,
  throw style, BBQ records, location, currency, unlocks, editor
  launchers, plus the helpers UnlockSvFlyLocations / UnlockSvBikeUpgrades
  / GetSvRotation / SetSvRotation / GetThrowStyleName / OpenFashion /
  OpenRaid and the SvFlyLocationBlockNames table.
- New: TrainerInfoSav9ZaSection (.razor + .razor.cs) — owns location,
  street name, currency, collect buttons, editor launchers, plus
  CollectZaTechnicalMachines / CollectZaColorfulScrews / GetZaStreetName
  / SetZaStreetName / OpenFashion / OpenDonut.
- TrainerInfoTab loses the two large SAV9SV / SAV9ZA inline blocks, the
  separate SAV9SV birthday block, the SV/ZA cases inside
  GetExtraCurrencies, and the SV/ZA helper methods now living in the
  sections.
- New CSS class .form-section-header — full-width grid item used to
  divide a .form-grid into logical sub-sections.
- _Imports gains the new Pkmds.Rcl.Components.MainTabPages.Trainer
  namespace so the sections can be referenced unqualified.
