+++ Pkmds.Rcl/_Imports.razor
@@ -30,6 +30,7 @@
 @using Pkmds.Rcl.Components.MainTabPages.Pokedex.Gen8La
 @using Pkmds.Rcl.Components.MainTabPages.Pokedex.Gen9Sv
 @using Pkmds.Rcl.Components.MainTabPages.Pokedex.Gen9Za
+@using Pkmds.Rcl.Components.MainTabPages.Trainer
 @using Pkmds.Rcl.Extensions
 @using Pkmds.Rcl.Services
 @using static GameInfoUtilities
